fixed: a few gcc warnings.  const char* in DVDAudio.cpp (second try)
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/linuxport/XBMC@12541 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/cores/VideoRenderers/LinuxRendererGL.cpp
@@ -753,7 +753,7 @@ void CLinuxRendererGL::SelectUpscalingMethod()
   
   // See if we're a candiate for upscaling.
   bool candidateForUpscaling = false;
-  if (upscale != SOFTWARE_UPSCALING_DISABLED && m_iSourceHeight < m_upscalingWidth && m_iSourceHeight < m_upscalingHeight)
+  if (upscale != SOFTWARE_UPSCALING_DISABLED && (int)m_iSourceHeight < m_upscalingWidth && (int)m_iSourceHeight < m_upscalingHeight)
   {
     CLog::Log(LOGWARNING, "Upscale: possible given resolution increase.");
     candidateForUpscaling = true;
